refactor(json-crdt): 💡 move proxy types one folder up
